% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pbmc3k.sce.final}
\alias{pbmc3k.sce.final}
\title{PBMC 3k Final}
\format{
A \sce object with the \code{pbmc3k.sce.final} dataset
}
\source{
\url{https://support.10xgenomics.com/single-cell-gene-expression/datasets/1.1.0/pbmc3k}

\url{https://satijalab.org/seurat/v3.0/pbmc3k_tutorial.html}

\url{http://bioconductor.org/books/3.17/OSCA.intro/analysis-overview.html}
}
\usage{
pbmc3k.sce.final
}
\description{
PBMC 3k Final
}
\seealso{

\code{\link{pbmc3k.sce}}
}
\concept{pbmc3ksce}
\keyword{datasets}
