% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pbmc3k.sce}
\alias{pbmc3k.sce}
\title{PBMC 3k}
\format{
A \sce object with the \code{pbmc3k.sce} dataset
}
\source{
\url{https://support.10xgenomics.com/single-cell-gene-expression/datasets/1.1.0/pbmc3k}
}
\usage{
pbmc3k.sce
}
\description{
PBMC 3k
}
\seealso{

\code{\link{pbmc3k.sce.final}}
}
\concept{pbmc3ksce}
\keyword{datasets}
