% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pbmc3k}
\alias{pbmc3k}
\title{pbmc3k}
\format{
A \sce object with the pbmc3k dataset
}
\source{
\url{https://support.10xgenomics.com/single-cell-gene-expression/datasets/1.1.0/pbmc3k}
}
\usage{
pbmc3k
}
\description{
pbmc3k
}
\keyword{datasets}
